# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nanaya",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token from /users/login"
      }
    }
  },
  "paths": {
...
    "/strip/{jobId}": {
      "get": {
        "summary": "Poll a strip job",
        "description": "Returns the status of a metadata-strip job. When done, includes the documentId.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "done",
                        "failed"
                      ]
                    },
                    "documentId": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "Semantic search over documents",
        "description": "Embeds the query string and returns documents ranked by cosine similarity.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query"
                ],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Natural language search query"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked list of matching documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "documentId": {
                            "type": "string"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "similarity": {
                            "type": "number",
                            "format": "float",
                            "minimum": -1,
                            "maximum": 1
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing query in request body"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Embedding or search failed"
          }
        }
      }
    },
    "/random": {
      "get": {
        "summary": "Get a random document",
        "description": "Returns metadata for a randomly selected document from the entire database.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The ID of a randomly selected document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documentId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "No documents found"
          }
        }
      }
    },
    "/my_documents": {
      "get": {
        "summary": "List your own documents",
        "description": "Returns metadata for all documents uploaded by the authenticated user.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of your documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "documentId": {
                        "type": "string"
                      },
                      "filename": {
                        "type": "string"
                      },
                      "uploadedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "processedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        }
      }
    },
    "/by_user/{userId}": {
      "get": {
        "summary": "List documents by user",
        "description": "Returns metadata for all documents uploaded by a specific user.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user whose documents to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents by the specified user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "documentId": {
                        "type": "string"
                      },
                      "filename": {
                        "type": "string"
                      },
                      "uploadedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "processedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        }
      }
    },
...
    "/{documentId}/download": {
      "get": {
        "summary": "Download a document",
        "description": "Returns the stripped DOCX file for any document.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the document to download"
          }
        ],
        "responses": {
          "200": {
            "description": "The DOCX file",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Document not found"
          }
        }
      }
    },
...
  },
...
}